+++ packages/jest-haste-map/src/__tests__/worker.test.js
@@ -11,7 +11,7 @@
 import * as path from 'path';
 import * as fs from 'graceful-fs';
 import H from '../constants';
-import {getSha1, process} from '../worker';
+import {extractMetadata, getSha1} from '../worker';
 
 jest.mock('graceful-fs', () => {
   const path = require('path');
@@ -59,14 +59,14 @@ jest.mock('graceful-fs', () => {
 
 const rootDir = '/project';
 
-describe('process', () => {
+describe('extractMetadata', () => {
   beforeEach(() => {
     jest.clearAllMocks();
   });
 
   it('parses JavaScript files and extracts module information', async () => {
     expect(
-      await process({
+      await extractMetadata({
         computeDependencies: true,
         filePath: path.join('/project', 'fruits', 'Pear.js'),
         rootDir,
@@ -76,7 +76,7 @@ describe('process', () => {
     });
 
     expect(
-      await process({
+      await extractMetadata({
         computeDependencies: true,
         filePath: path.join('/project', 'fruits', 'Strawberry.js'),
         rootDir,
@@ -88,7 +88,7 @@ describe('process', () => {
 
   it('accepts a custom dependency extractor', async () => {
     expect(
-      await process({
+      await extractMetadata({
         computeDependencies: true,
         dependencyExtractor: path.join(__dirname, 'dependencyExtractor.js'),
         filePath: path.join('/project', 'fruits', 'Pear.js'),
@@ -101,7 +101,7 @@ describe('process', () => {
 
   it('delegates to hasteImplModulePath for getting the id', async () => {
     expect(
-      await process({
+      await extractMetadata({
         computeDependencies: true,
         filePath: path.join('/project', 'fruits', 'Pear.js'),
         hasteImplModulePath: require.resolve('./haste_impl.js'),
@@ -114,7 +114,7 @@ describe('process', () => {
     });
 
     expect(
-      await process({
+      await extractMetadata({
         computeDependencies: true,
         filePath: path.join('/project', 'fruits', 'Strawberry.js'),
         rootDir,
@@ -128,7 +128,7 @@ describe('process', () => {
 
   it('parses package.json files as haste packages', async () => {
     expect(
-      await process({
+      await extractMetadata({
         computeDependencies: true,
         filePath: path.join('/project', 'package.json'),
         rootDir,
@@ -144,7 +144,11 @@ describe('process', () => {
     let error = null;
 
     try {
-      await process({computeDependencies: true, filePath: '/kiwi.js', rootDir});
+      await extractMetadata({
+        computeDependencies: true,
+        filePath: '/kiwi.js',
+        rootDir,
+      });
     } catch (err) {
       error = err;
     }
@@ -192,7 +196,7 @@ describe('process', () => {
 
   it('avoids computing dependencies if not requested and Haste does not need it', async () => {
     expect(
-      await process({
+      await extractMetadata({
         computeDependencies: false,
         filePath: path.join('/project', 'fruits', 'Pear.js'),
         hasteImplModulePath: path.resolve(__dirname, 'haste_impl.js'),