+++ packages/jest-haste-map/src/types.ts
@@ -12,7 +12,7 @@ import type ModuleMap from './ModuleMap';
 
 export type IgnoreMatcher = (item: string) => boolean;
 
-export type WorkerMessage = {
+export type ExtractMetadataDefinition = {
   computeDependencies: boolean;
   computeSha1: boolean;
   dependencyExtractor?: string | null;
@@ -21,22 +21,35 @@ export type WorkerMessage = {
   hasteImplModulePath?: string;
 };
 
-export type WorkerMetadata = {
+export type ExtractedFileMetaData = {
   dependencies: Array<string> | undefined | null;
   id: string | undefined | null;
   module: ModuleMetaData | undefined | null;
   sha1: string | undefined | null;
 };
 
-export type WorkerInterface = {
-  process(data: WorkerMessage): Promise<WorkerMetadata>;
-  getSha1(data: WorkerMessage): Promise<WorkerMetadata>;
-  end(): void;
+export type MetadataExtractorOptions = {
+  forceInBand: boolean;
 };
 
-export type WorkerFactory = (options: {
-  forceInBand: boolean;
-}) => WorkerInterface;
+export interface MetadataExtractor {
+  /**
+   * Called before either `extractMetadata` or `getSha1` is called.
+   */
+  setup(options: MetadataExtractorOptions): void;
+
+  extractMetadata(
+    data: ExtractMetadataDefinition,
+  ): Promise<ExtractedFileMetaData>;
+
+  getSha1(data: ExtractMetadataDefinition): Promise<ExtractedFileMetaData>;
+
+  /**
+   * Called when the extractor is no longer in use to allow releasing resources.
+   * Called even if `setup` was never called.
+   */
+  end(): void;
+}
 
 export type CrawlerOptions = {
   computeSha1: boolean;